+++ pgsql84.c
@@ -139,7 +139,7 @@
  * by using the final values of both b and c.  b is perhaps a little less
  * well mixed than c, however.
  */
-uint64_t hlib_pgsql84(const void *data, uint32_t keylen, uint32_t seed)
+uint64_t hlib_pgsql84(const void *data, size_t keylen, uint64_t seed)
 {
 	const unsigned char *k = (unsigned char *)data;
 	register uint32 a,
